microservices/chatbot/src/pages/chatbot/requirements.txt
google-generativeai
flask
flask-cors